Move dispatch events to their own module
These events are used by both the actor module and the top-level module,
and we want to expose them publicly only in one place (the top-level
module), so we don't confuse the users about where they should import
symbols from.
